requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.4
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
psycopg2-binary==2.9.10
sqlparse==0.5.3
